moved authentication to a separate file
