+++ app/src/Authentication.tsx
@@ -1,0 +1,57 @@
+// This was copied from this tutorial: https://medium.com/geekculture/firebase-auth-with-react-and-typescript-abeebcd7940a
+import React, { useContext, useRef } from "react";
+import { AuthContext } from "./context/AuthContext";
+import { auth } from "./firebaseSetup";
+import { getAuth, signInWithRedirect, GoogleAuthProvider } from "firebase/auth";
+
+function App() {
+  const user = useContext(AuthContext);
+
+  // Email and password sign in
+  const emailRef = useRef<HTMLInputElement>(null);
+  const passwordRef = useRef<HTMLInputElement>(null);
+
+  // Google authentication
+  const provider = new GoogleAuthProvider();
+
+  const googleSignIn = async () => {
+
+    // Code from Firebase documentation
+    const auth = getAuth();
+    signInWithRedirect(auth, provider);
+    
+  }
+
+  const createAccount = async () => {
+    try {
+      await auth.createUserWithEmailAndPassword(
+        emailRef.current!.value,
+        passwordRef.current!.value
+      );
+    } catch (error) {
+      console.error(error);
+    }
+  };
+
+  const signIn = async () => {
+    try {
+      await auth.signInWithEmailAndPassword(
+        emailRef.current!.value,
+        passwordRef.current!.value
+      );
+    } catch (error) {
+      console.error(error);
+    }
+  };
+
+  const signOut = async () => {
+    await auth.signOut();
+  };
+
+  return (
+    <>
+    </>
+  );
+}
+
+export default App;
